menu implementation(5):20140310

+++ usr/local/opendomo/vr/autodetect.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-play BEEP1.WAV
-rec recording.flac rate 8k silence 1 0.1 1% 1 3.0 3% &
+#play BEEP1.WAV
+rec recording.flac rate 8k silence 1 0.1 3% 1 3.0 3% &
 p=$!
 sleep 1
 until [ "$var1" != "$var2" ]; do
@@ -16,8 +16,12 @@ until [ "$var1" == "$var2" ]; do
 done
 echo "Silence Detected"
 
-kill $p
-play recording.flac
+if [ `ps -eaf | grep -c $p` -gt 1 ] 
+#grep is 1...minimum value is 1
+then
+	kill $p
+fi
+
 RESULT="$(./send_speech recording.flac)"
 ./recognize "$RESULT"
